added authentication

+++ backend/src/main/java/com/fashionstore/services/AddressService.java
@@ -6,6 +6,8 @@ import com.fashionstore.exceptions.NotFoundException;
 import org.springframework.stereotype.Service;
 import com.fashionstore.models.Address;
 import com.fashionstore.repositories.AddressRepository;
+import com.fashionstore.repositories.UserRepository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -14,11 +16,14 @@ import java.util.stream.Collectors;
 @Service
 public class AddressService {
     private final AddressRepository addressRepository;
+    private final UserRepository userRepository;
 
-    public AddressService(AddressRepository addressRepository) {
+    public AddressService(AddressRepository addressRepository, UserRepository userRepository) {
         this.addressRepository = addressRepository;
+        this.userRepository = userRepository;
     }
 
+    @Transactional
     public AddressResponse addAddress(AddressRequest addressRequest) {
         Address newAddress = new Address();
         newAddress.setCity(addressRequest.getCity());
@@ -26,15 +31,19 @@ public class AddressService {
         newAddress.setCountry(addressRequest.getCountry());
         newAddress.setRegion(addressRequest.getRegion());
         newAddress.setPostalCode(addressRequest.getPostalCode());
+        newAddress.setUser(userRepository.findById(addressRequest.getUserId())
+                .orElseThrow(() -> new NotFoundException("User", addressRequest.getUserId())));
         Address savedAddress = addressRepository.save(newAddress);
         return AddressResponse.from(savedAddress);
     }
 
+    @Transactional(readOnly = true)
     public AddressResponse getAddressById(Long id) {
         Optional<Address> address = addressRepository.findById(id);
         return  address.map(AddressResponse::from).orElseThrow(() -> new NotFoundException("Address", id));
     }
 
+    @Transactional(readOnly = true)
     public List<AddressResponse> getAllAddresses() {
         return addressRepository.findAll()
                 .stream()
@@ -42,6 +51,7 @@ public class AddressService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public AddressResponse updateAddress(Long id, AddressRequest addressRequest){
         Optional<Address> address = addressRepository.findById(id);
         if (address.isPresent()) {
@@ -51,13 +61,19 @@ public class AddressService {
             updatedAddress.setCountry(addressRequest.getCountry());
             updatedAddress.setRegion(addressRequest.getRegion());
             updatedAddress.setPostalCode(addressRequest.getPostalCode());
+            updatedAddress.setUser(userRepository.findById(addressRequest.getUserId())
+                    .orElseThrow(() -> new NotFoundException("User", addressRequest.getUserId())));
             Address savedAddress = addressRepository.save(updatedAddress);
             return AddressResponse.from(savedAddress);
         }
         throw new NotFoundException("Address", id);
     }
 
+    @Transactional
     public void deleteAddress(Long id){
+        if (!addressRepository.existsById(id)) {
+            throw new NotFoundException("Address", id);
+        }
         addressRepository.deleteById(id);
     }
 }
